is24 - is25 - is26 - removed validations of inside of controllers
