+++ src/app/controllers/SessionController.js
@@ -1,4 +1,3 @@
-import * as Yup from 'yup';
 import jwt from 'jsonwebtoken';
 
 import User from '../schemas/User';
@@ -6,19 +5,6 @@ import authConfig from '../../config/auth';
 
 class SessionController {
   async store(req, res) {
-    const schema = Yup.object().shape({
-      email: Yup.string()
-        .email()
-        .required(),
-      password: Yup.string().required(),
-    });
-
-    if (!(await schema.isValid(req.body))) {
-      return res
-        .status(400)
-        .json({ error: 'Body of request out of expected format!' });
-    }
-
     const { email, password } = req.body;
 
     const user = await User.findOne({ email });
