+++ src/app/validators/SessionStore.js
@@ -1,0 +1,23 @@
+import * as Yup from 'yup';
+
+export default async (req, res, next) => {
+  try {
+    const schema = Yup.object().shape({
+      email: Yup.string()
+        .email()
+        .required(),
+      password: Yup.string()
+        .required()
+        .min(6),
+    });
+
+    await schema.validate(req.body, { abortEarly: false });
+
+    return next();
+  } catch (err) {
+    return res.status(400).json({
+      error: 'Validation fails, body of request out of expected format!',
+      messages: err.inner,
+    });
+  }
+};
